backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
pydantic
python-dotenv
